Day09 - Solve Part 02
As the LinkedList and index lookup implementation was way too slow (I
tried and couldn't even get it to finish) I needed something different.

Ended up using ArrayDeque, as it seemed to have a quite good performance,
and I used rotation to be able to just use addFirst and addLast to not
need any (slow) index lookups and removals in the middle of the queue
anymore.

After I finally wrapped my head around rotating correctly, this
implementation is _a lot_ faster than the previous one.

Finally, I needed to replace the Int scores with Long, as the scores do
get quite big and I ran into an Integer overflow.